Changed from button to command
